refactor: split billing cleanup into invoice-only module, simplify handleInvoiceEvent
- billing-cleanup.js now only contains invoice refund helpers
  (getStripeObjectId, refundDeletedAccountInvoice) keeping it small
  and below the CodeScene health threshold
- Checkout cleanup functions remain in billing.js
- Extract isInvoiceForDeletedAccount helper to simplify handleInvoiceEvent
  and reduce its branch complexity below threshold
