+++ server/billing.js
@@ -8,12 +8,7 @@ const { getAdminAuth, getAdminDb, hasFirebaseAdminConfig } = require('./firebase
 const { getBaseUrl, getBillingRuntimeConfig, getPublicBillingConfig } = require('./billing-config');
 const { recordBillingMetricEvent } = require('./metrics');
 const { deleteStripeCustomer, isAccountDeletionBlocked, isStripeCustomerDeletionBlocked } = require('./account-lifecycle');
-const {
-  cleanupBlockedCheckoutSession,
-  getCheckoutRefundTarget,
-  getStripeObjectId,
-  refundDeletedAccountInvoice,
-} = require('./billing-cleanup');
+const { getStripeObjectId, refundDeletedAccountInvoice } = require('./billing-cleanup');
 
 let stripeClient = null;
 const RATE_LIMIT_WINDOW_MS = 60 * 1000;
@@ -408,6 +403,108 @@ const createSubscriptionEntitlementWrite = (subscription) => {
   };
 };
 
+/** Returns the first usable Stripe ID from a list of expanded objects or plain IDs. */
+const getFirstStripeObjectId = (values) => values.map(getStripeObjectId).find(Boolean) || '';
+
+/** Finds the payment intent across Checkout and legacy/current invoice shapes. */
+const getCheckoutPaymentIntentId = (session, invoice, payments) =>
+  getFirstStripeObjectId([
+    session.payment_intent,
+    invoice?.payment_intent,
+    ...payments.map((payment) => payment?.payment?.payment_intent),
+  ]);
+
+/** Finds a charge-only payment across current and legacy invoice shapes. */
+const getCheckoutChargeId = (invoice, payments) =>
+  getFirstStripeObjectId([
+    ...payments.map((payment) => payment?.payment?.charge),
+    invoice?.charge,
+  ]);
+
+/** Finds the initial Checkout payment across Stripe's legacy and current invoice shapes. */
+const getCheckoutRefundTarget = (session, subscription) => {
+  const invoice = subscription?.latest_invoice;
+  const payments = invoice?.payments?.data || [];
+  const paymentIntent = getCheckoutPaymentIntentId(session, invoice, payments);
+  if (paymentIntent) return { payment_intent: paymentIntent };
+
+  const charge = getCheckoutChargeId(invoice, payments);
+  return charge ? { charge } : null;
+};
+
+/** Returns the UID only when this Checkout belongs to an account blocked from writes. */
+const getBlockedCheckoutUid = async (session, canWrite) => {
+  const uid = session.client_reference_id || session.metadata?.uid || '';
+  if (!uid) return null;
+  return (await canWrite(uid)) ? null : uid;
+};
+
+/** Retrieves the expanded first invoice for a subscription-mode Checkout session. */
+const getCheckoutSubscription = async (stripe, session) => {
+  const subscriptionId = getStripeObjectId(session.subscription);
+  if (!subscriptionId) return null;
+  try {
+    return await stripe.subscriptions.retrieve(subscriptionId, { expand: ['latest_invoice'] });
+  } catch (error) {
+    if (error?.code === 'resource_missing') return null;
+    throw error;
+  }
+};
+
+/** Resolves a refundable payment using both legacy invoice fields and the current payments API. */
+const resolveCheckoutRefundTarget = async (stripe, session, subscription) => {
+  const embeddedTarget = getCheckoutRefundTarget(session, subscription);
+  if (embeddedTarget) return embeddedTarget;
+
+  const invoiceId = getStripeObjectId(subscription?.latest_invoice);
+  if (!invoiceId || !stripe.invoicePayments?.list) return null;
+  const invoicePayments = await stripe.invoicePayments.list({ invoice: invoiceId, limit: 10 });
+  return getCheckoutRefundTarget(session, {
+    latest_invoice: { payments: { data: invoicePayments.data || [] } },
+  });
+};
+
+/** Prevents a paid late Checkout from being cleaned up without its required refund. */
+const ensurePaidCheckoutHasRefundTarget = (session, refundTarget) => {
+  if (session.payment_status !== 'paid' || refundTarget) return;
+  throw new Error(`Unable to identify payment for deleted-account Checkout session ${session.id}.`);
+};
+
+/** Creates the refund only when Checkout actually produced a refundable payment. */
+const refundBlockedCheckout = async (stripe, session, refundTarget) => {
+  if (!refundTarget) return;
+  await stripe.refunds.create(refundTarget, {
+    idempotencyKey: `account-deletion-checkout-refund:${session.id}`,
+  });
+};
+
+/**
+ * Refunds a Checkout payment that completed after account deletion, then removes
+ * the newly created Stripe customer. Throwing keeps the webhook retryable.
+ */
+const cleanupBlockedCheckoutSession = async (
+  session,
+  { stripe = getStripeClient(), canWrite = canWriteEntitlementForUid } = {}
+) => {
+  const uid = await getBlockedCheckoutUid(session, canWrite);
+  if (!uid) return false;
+  if (!stripe) throw new Error('Stripe is not configured for late Checkout cleanup.');
+
+  if (session.payment_status === 'unpaid') return false;
+
+  const subscription = await getCheckoutSubscription(stripe, session);
+  const refundTarget = await resolveCheckoutRefundTarget(stripe, session, subscription);
+  ensurePaidCheckoutHasRefundTarget(session, refundTarget);
+  await refundBlockedCheckout(stripe, session, refundTarget);
+
+  await deleteStripeCustomer({ stripe, customerId: getStripeObjectId(session.customer) });
+  console.warn('[billing] checkout:refunded-deleted-user', {
+    uid,
+    sessionId: redactIdentifier(session.id),
+  });
+  return true;
+};
+
 /** Records billing metric events with safe error handling. */
 const recordBillingMetricEventSafely = async (eventType) => {
   try {
@@ -497,6 +594,14 @@ const resolveInvoiceUid = (subscription, invoice) =>
   invoice.subscription_details?.metadata?.uid ||
   '';
 
+/** Checks whether an invoice belongs to a deleted account. */
+const isInvoiceForDeletedAccount = async (subscription, invoice, customerId) => {
+  const uid = resolveInvoiceUid(subscription, invoice);
+  return uid
+    ? isAccountDeletionBlocked(getAdminDb(), uid)
+    : isStripeCustomerDeletionBlocked(getAdminDb(), customerId);
+};
+
 /** Applies invoice notifications from the current authoritative subscription state. */
 const handleInvoiceEvent = async (event, stripe) => {
   const subscription = await resolveAuthoritativeSubscription(stripe, event);
@@ -508,18 +613,10 @@ const handleInvoiceEvent = async (event, stripe) => {
   const invoice = event.data.object;
   const invoiceStripe = stripe || getStripeClient();
   const invoiceCustomerId = getStripeObjectId(invoice.customer);
-  if (result?.reason === 'deleted-user') {
+  const shouldRefund = result?.reason === 'deleted-user' ||
+    (result?.reason === 'missing-target' && await isInvoiceForDeletedAccount(subscription, invoice, invoiceCustomerId));
+  if (shouldRefund) {
     await refundDeletedAccountInvoice(invoice, { stripe: invoiceStripe, customerId: invoiceCustomerId });
-    return;
-  }
-  if (result?.reason === 'missing-target') {
-    const uid = resolveInvoiceUid(subscription, invoice);
-    const isDeleted = uid
-      ? await isAccountDeletionBlocked(getAdminDb(), uid)
-      : await isStripeCustomerDeletionBlocked(getAdminDb(), invoiceCustomerId);
-    if (isDeleted) {
-      await refundDeletedAccountInvoice(invoice, { stripe: invoiceStripe, customerId: invoiceCustomerId });
-    }
   }
 };
